LayoutTests/storage/indexeddb/objectStore-required-arguments-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB object store required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo');
PASS objectStore.put(); threw exception TypeError: Not enough arguments.
PASS objectStore.add(); threw exception TypeError: Not enough arguments.
PASS objectStore.delete(); threw exception TypeError: Not enough arguments.
PASS objectStore.get(); threw exception TypeError: Not enough arguments.
PASS objectStore.createIndex(); threw exception TypeError: Not enough arguments.
PASS objectStore.createIndex('foo'); threw exception TypeError: Not enough arguments.
PASS objectStore.index(); threw exception TypeError: Not enough arguments.
PASS objectStore.deleteIndex(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
